Add files via upload
Starting out - established the base structures and am working on how to present the maze. Currently, the plan is to use a 2D int vector where each number represents a part of the maze. 

To-do:
Maze visuals - Each part of the maze will be a 5x5 block that will be outputted. The area that will be visible isn't the full maze but only up to 4 blocks away. This is so Alice can actually be drawn out in Unicode characters and have an animation.
Mini map - A way for the player to see the full map
Sprites - Work on Alice's sprite and the in-maze default sprite of the monsters.
Monsters - Work out which monster will appear during an encounter, along with their attacks. Enums and/or const arrays would work best.
Battle interface - Work out how battles will work. It will be a turn-based system. Who attacks per round first will either be determined by randomisation, or it will always be Alice. Add a default attack, an option for Alice to equip a weapon (and weapons-but-actually-items)
Weapons - Add weapons. Allow Alice to access weapons while navigating the maze. Decide if Alice should be able to waste a weapon out of battle? Or if, instead, only certain weapons can be used.
Animation - Alice's and the monster's animation will have 3 frames - walking 1, walking 2, walking 3, along with an idle state. Walking 1 and 2 will depend on the direction. Say, if Alice is walking to the left, there will be two different frames for each foot and an in-between state where both feet are on the ground. If Alice is walking forward, the foot raised will change and Alice herself will move her head up and down to show movement.
Starting Screen - Cool animation goes boom
Other things my brain can't remember currently but will end up so vitally important that the program won't run without it

Wishful additions: 
Graphics - no idea
Colours - Possible but last addition
Images - sorta a lesser version of graphics but more doable potentially. Each block will be a png image instead.
No flickering during output update - Happens due to system("CLS"), so I need to find another way to clear the output during each frame. Priority from wishful additions since it's not actually a pain.